requirements.txt
# API Requirements
flask>=2.3.3,<3.0
gunicorn>=21.2.0,<22.0
psycopg2-binary>=2.9.7,<3.0
google-cloud-pubsub>=2.18.4,<3.0
python-json-logger>=2.0.7,<3.0
python-dotenv>=1.0.0,<2.0
requests>=2.31.0,<3.0

# Development dependencies
black>=23.7.0,<24.0
flake8>=6.1.0,<7.0
isort>=5.12.0,<6.0
mypy>=1.5.1,<2.0
pytest>=7.4.0,<8.0
pytest-cov>=4.1.0,<5.0

# Type checking dependencies
types-flask
types-psycopg2

# Additional test/dev libraries
pytest-mock>=3.6.1,<4.0
requests-mock>=1.9.3,<2.0
pytest>=7.4,<8.0